15/4 commit

+++ src/main/java/com/swiftdroid/posterhouse/controller/ShoppingCartController.java
@@ -20,15 +20,18 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.swiftdroid.posterhouse.config.CustomOAuth2User;
 import com.swiftdroid.posterhouse.model.CartItem;
+import com.swiftdroid.posterhouse.model.CartItemToImage;
 import com.swiftdroid.posterhouse.model.Product;
 import com.swiftdroid.posterhouse.model.ProductConfig;
 import com.swiftdroid.posterhouse.model.ShoppingCart;
 import com.swiftdroid.posterhouse.model.User;
+import com.swiftdroid.posterhouse.repo.CartItemImageRepository;
 import com.swiftdroid.posterhouse.service.CartItemService;
 import com.swiftdroid.posterhouse.service.ProductConfigService;
 import com.swiftdroid.posterhouse.service.ProductService;
 import com.swiftdroid.posterhouse.service.ShoppingCartService;
 import com.swiftdroid.posterhouse.service.UserService;
+import com.swiftdroid.posterhouse.serviceimpl.CartToImageService;
 
 @Controller
 public class ShoppingCartController {
@@ -42,6 +45,11 @@ public class ShoppingCartController {
 	/*
 	 * @Autowired private BookService bookService;
 	 */
+	@Autowired
+	private CartItemImageRepository cartItemImageImpl;
+
+	@Autowired
+	private CartToImageService cartToImageService;
 
 	@Autowired
 	private ShoppingCartService shoppingCartService;
@@ -67,9 +75,9 @@ public class ShoppingCartController {
 
 			ShoppingCart shoppingCart = user.getShoppingCart();
 
-			List<CartItem> cartItemList = cartItemService.findByShoppingCart(shoppingCart);
-
 			shoppingCartService.updateShoppingCart(shoppingCart);
+
+			List<CartItem> cartItemList = cartItemService.findByShoppingCart(shoppingCart);
 
 			model.addAttribute("cartItemList", cartItemList);
 			model.addAttribute("shoppingCart", shoppingCart);
@@ -82,10 +90,9 @@ public class ShoppingCartController {
 	@SuppressWarnings("finally")
 	@RequestMapping("/addItem")
 	public String addItem(@ModelAttribute("qty") String qty,
-			@ModelAttribute("productConfig") ProductConfig productConfig,
-
-			Model model, Principal principal, Authentication authentication,
-			@RequestParam("file") MultipartFile[] imageMultipart) throws MultipartException {
+			@ModelAttribute("productConfig") ProductConfig productConfig, Model model, Principal principal,
+			Authentication authentication, @RequestParam("file") MultipartFile[] imageMultipart)
+			throws MultipartException {
 
 		if (productConfig.getId() == null) {
 			System.out.println("ProductConfig Id Null  Controller ShoppingCartController :: Method addItem :: ");
@@ -112,114 +119,127 @@ public class ShoppingCartController {
 			System.out.println(
 					"*************************************************************22*********************************************");
 
-			System.out.println("yyyy" + imageMultipart.length);
+			System.out.println("yyyy  " + imageMultipart.length);
+			System.out.println(product.getMaximumQuantity());
+			int quantity = imageMultipart.length;
 
-			if (imageMultipart.length > product.getMaximumQuantity()) {
+			int maximumQuantity = product.getMaximumQuantity();
+
+			if (quantity > maximumQuantity) {
 
 				model.addAttribute("notEnoughStock", true);
 
 				return "forward:/productDetail?id=" + product.getId();
 			}
 			try {
+
 				List<CartItem> cartItemList = user.getShoppingCart().getCartItemList();
-				System.out.println();
-				if(cartItemList.size()!=0) {
-				for (CartItem cartItem : cartItemList) {
+				System.out.println("@@@@@@@@@@@ 1 @@@@@@@");
+				if (cartItemList.size() != 0) {
+					for (CartItem cartItem : cartItemList) {
+						System.out.println("@@@@@@@@@@@ 2 @@@@@@@");
 
-					if (cartItem.getProduct().equals(product)) {
-						String path = null;
-						int count = cartItem.getQty();
-						for (MultipartFile multipartFile : imageMultipart) {
+						if (cartItem.getProduct().equals(product)) {
+							System.out.println("@@@@@@@@@@@ 3 @@@@@@@");
 
-							byte[] bytes = multipartFile.getBytes();
-							count++;
-							String name = user.getShoppingCart().getId() + "_" + product.getId() + "_" + count
-									+ "_.png";
+							for (MultipartFile multipartFile : imageMultipart) {
+								System.out.println("@@@@@@@@@@@ 3 loop @@@@@@@");
 
-							File file = new File("src/main/resources/static/img/user/userproductImage/" + name);
-							FileOutputStream out = new FileOutputStream(file);
-							BufferedOutputStream strem = new BufferedOutputStream(out);
-							strem.write(bytes);
-							path = "/img/user/userproductImage/" + name;
-							strem.close();
+								byte[] bytes = multipartFile.getBytes();
+
+								CartItemToImage cartItemImage = new CartItemToImage();
+								cartItemImage.setCartItem(cartItem);
+								cartItemImage = cartItemImageImpl.save(cartItemImage);
+								String name = user.getShoppingCart().getId() + "_" + product.getId() + "_"
+										+ cartItem.getId() + "_" + cartItemImage.getId() + "_.png";
+								cartItemImage.setImgPath(name);
+								// String id=user.getShoppingCart().getId() + "_" + product.getId() + "_.png"
+								File file = new File("src/main/resources/static/img/user/userproductImage/" + name);
+								FileOutputStream out = new FileOutputStream(file);
+								BufferedOutputStream strem = new BufferedOutputStream(out);
+								cartItemImageImpl.save(cartItemImage);
+								strem.write(bytes);
+								strem.close();
+							}
+
+							System.out.println(
+									"*************************************************************33*********************************************");
+							System.out.println("productConfig::::" + productConfig);
+
+							System.out.println("@@@@@@@@@@@ 4 @@@@@@@");
+
+							CartItem cartItems = cartItemService.addProductToCartItem("", productConfig, product, user,
+									imageMultipart.length);
+
+							model.addAttribute("addBookSuccess", true);
+
+							return "forward:/productDetail?id=" + product.getId();
+
+						} else {
+							CartItem cartItems = cartItemService.addProductToCartItem("", productConfig, product, user,
+									imageMultipart.length);
+
+							System.out.println("@@@@@@@@@@@ 5 @@@@@@@");
+
+							for (MultipartFile multipartFile : imageMultipart) {
+
+								byte[] bytes = multipartFile.getBytes();
+
+								CartItemToImage cartItemImage = new CartItemToImage();
+								cartItemImage.setCartItem(cartItems);
+								cartItemImage = cartItemImageImpl.save(cartItemImage);
+								String name = user.getShoppingCart().getId() + "_" + product.getId() + "_"
+										+ cartItems.getId() + "_" + cartItemImage.getId() + "_.png";
+								cartItemImage.setImgPath(name);
+								// String id=user.getShoppingCart().getId() + "_" + product.getId() + "_.png"
+								File file = new File("src/main/resources/static/img/user/userproductImage/" + name);
+								FileOutputStream out = new FileOutputStream(file);
+								BufferedOutputStream strem = new BufferedOutputStream(out);
+								cartItemImageImpl.save(cartItemImage);
+								strem.write(bytes);
+								strem.close();
+							}
+							model.addAttribute("addBookSuccess", true);
+							System.out.println("@@@@@@@@@@@ 3@@@@@@@");
+							return "forward:/productDetail?id=" + product.getId();
 						}
-						System.out.println(
-								"*************************************************************33*********************************************");
-						System.out.println("productConfig::::" + productConfig);
-
-						CartItem cartItems = cartItemService.addProductToCartItem(path, productConfig, product, user,
-								count);
-
-						model.addAttribute("addBookSuccess", true);
-
-						return "forward:/productDetail?id=" + product.getId();
-					} else {
-						String newpath = null;
-						int newcount = 0;
-
-						for (MultipartFile multipartFile : imageMultipart) {
-
-							byte[] bytes = multipartFile.getBytes();
-							newcount++;
-							String name = user.getShoppingCart().getId() + "_" + product.getId() + "_" + newcount
-									+ "_.png";
-
-							File file = new File("src/main/resources/static/img/user/userproductImage/" + name);
-							FileOutputStream out = new FileOutputStream(file);
-							BufferedOutputStream strem = new BufferedOutputStream(out);
-							strem.write(bytes);
-							newpath = "/img/user/userproductImage/" + name;
-							strem.close();
-						}
-						System.out.println(
-								"*************************************************************33*********************************************");
-						System.out.println("productConfig::::" + productConfig);
-
-						CartItem cartItems = cartItemService.addProductToCartItem(newpath, productConfig, product, user,
-								newcount);
-
-						model.addAttribute("addBookSuccess", true);
-
-						return "forward:/productDetail?id=" + product.getId();
 					}
+				} else {
 
-				}
-			}
-				else {
-					String newpath = null;
-					int newcount = 0;
+					CartItem cartItems = cartItemService.addProductToCartItem("", productConfig, product, user,
+							imageMultipart.length);
 
 					for (MultipartFile multipartFile : imageMultipart) {
 
 						byte[] bytes = multipartFile.getBytes();
-						newcount++;
-						String name = user.getShoppingCart().getId() + "_" + product.getId() + "_" + newcount
-								+ "_.png";
 
+						CartItemToImage cartItemImage = new CartItemToImage();
+						cartItemImage.setCartItem(cartItems);
+						cartItemImage = cartItemImageImpl.save(cartItemImage);
+						String name = user.getShoppingCart().getId() + "_" + product.getId() + "_" + cartItems.getId()
+								+ "_" + cartItemImage.getId() + "_.png";
+						cartItemImage.setImgPath(name);
+						// String id=user.getShoppingCart().getId() + "_" + product.getId() + "_.png"
 						File file = new File("src/main/resources/static/img/user/userproductImage/" + name);
 						FileOutputStream out = new FileOutputStream(file);
 						BufferedOutputStream strem = new BufferedOutputStream(out);
+						cartItemImageImpl.save(cartItemImage);
 						strem.write(bytes);
-						newpath = "/img/user/userproductImage/" + name;
 						strem.close();
 					}
-					System.out.println(
-							"*************************************************************33*********************************************");
-					System.out.println("productConfig::::" + productConfig);
-
-					CartItem cartItems = cartItemService.addProductToCartItem(newpath, productConfig, product, user,
-							newcount);
-
 					model.addAttribute("addBookSuccess", true);
-
 					return "forward:/productDetail?id=" + product.getId();
+
 				}
-			}catch (Exception e) {
-				e.printStackTrace();
 			}
 
+			catch (Exception e) {
+				System.out.println(e.getMessage());
+
+			}
+
+			return "forward:/productDetail?id=" + product.getId();
 		}
-		return qty;
 	}
 
 	@RequestMapping("/updateCartItem")
@@ -249,8 +269,11 @@ public class ShoppingCartController {
 
 	@SuppressWarnings("finally")
 	@RequestMapping("/removeItem")
-	public String removeItem(@RequestParam("id") Long id, Authentication authentication, Principal principal) {
+	public String removeItem(@RequestParam("id") Long id, @RequestParam("count") Long imageId,
+			Authentication authentication, Principal principal, Model model) {
 		System.out.println("id  ::  " + id);
+		System.out.println("id  ::  " + imageId);
+
 		if (id == null) {
 
 			System.out.println("id  null In ShoppingCartController   Method:: removeItem  ::  ");
@@ -267,13 +290,25 @@ public class ShoppingCartController {
 			} catch (Exception e) {
 				user = userService.findByUsername(principal.getName());
 			} finally {
-				CartItem cartItem = cartItemService.findById(id);
-				Product product = cartItem.getProduct();
-				String name = user.getId() + "_" + product.getId() + ".png";
-				Files.delete(Paths.get("src/main/resources/static/img/user/userproductImage/" + name));
-				cartItemService.removeCartItem(cartItemService.findById(id));
 
+				CartItem cartItem = cartItemService.findById(id);
+
+				if (cartItem.getQty() == 0) {
+					model.addAttribute("qtyMsg", true);
+					cartItemService.removeCartItem(cartItemService.findById(id));
+					return "forward:/cart";
+				}
+
+				Product product = cartItem.getProduct();
+				ShoppingCart shoppingCart = user.getShoppingCart();
+				String name = user.getShoppingCart().getId() + "_" + product.getId() + "_" + cartItem.getId() + "_"
+						+ imageId + "_.png";
+				Files.delete(Paths.get("src/main/resources/static/img/user/userproductImage/" + name));
+				cartToImageService.deleteCartImageByCartIdAndCartImageId(cartItem, imageId);
+				cartItem.setQty(cartItem.getQty() - 1);
+				cartItemService.saveCart(cartItem);
 				return "forward:/cart";
+
 			}
 		} catch (Exception e) {
 			System.out.println("id  null In ShoppingCartController   Method:: removeItem  ::  " + e.getMessage());
